chore: rename asset handler source files
